+++ .github/workflows/test-export-wpt-changes.yml
@@ -2,7 +2,7 @@ name: WPT exporter test
 on:
   pull_request:
     branches: ["**"]
-    paths: ["etc/ci/upstream-wpt-changes/**"]
+    paths: ["python/wpt/exporter/**"]
 
 jobs:
   test:
@@ -12,7 +12,7 @@ jobs:
       - uses: actions/checkout@v3
       - name: Install dependencies
         run: |
-          python3 -m pip install --upgrade -r etc/ci/upstream-wpt-changes/requirements-dev.txt
+          python3 -m pip install --upgrade -r python/wpt/requirements-dev.txt
       - name: Running tests
         run: |
-          python3 etc/ci/upstream-wpt-changes/test.py
+          python3 python/wpt/test.py
