+++ python/wpt/exporter/common.py
