+++ python/wpt/tests/move-into-wpt.diff
