+++ python/wpt/tests/move-out-of-wpt.diff
