+++ python/wpt/tests/servo-mock/tests/wpt/web-platform-tests/fetch/api/redirect/redirect-location.html
