+++ python/wpt/tests/wpt-mock/fetch/api/redirect/redirect-location.html
