+++ servo-tidy.toml
@@ -136,11 +136,9 @@ directories = [
   "./support/magicleap/Servo2D/code/src.gen",
   "./support/magicleap/Servo2D/pipeline",
   "./tests/wpt/harness",
-  "./tests/wpt/update",
   "./tests/wpt/web-platform-tests",
   "./tests/wpt/mozilla/tests/mozilla/referrer-policy",
   "./tests/wpt/mozilla/tests/webgl",
-  "./tests/wpt/sync",
   "./python/tidy/servo_tidy_tests",
   "./components/script/dom/bindings/codegen/parser",
   "./components/script/dom/bindings/codegen/ply",
