+++ tests/wpt/README.md
@@ -9,7 +9,6 @@ In particular, this folder contains:
 
 * `config.ini`: some configuration for the web-platform-tests.
 * `include.ini`: the subset of web-platform-tests we currently run.
-* `servowpt.py`: run the web-platform-tests in Servo.
 * `web-platform-tests`: copy of the web-platform-tests.
 * `metadata`: expected failures for the web-platform-tests we run.
 * `mozilla`: web-platform-tests that cannot be upstreamed.
@@ -78,12 +77,6 @@ testharnessreport.js may have been installed incorrectly (see
 [**Running the tests manually**](#running-the-tests-manually)
 for more details).
 
-Running the tests without mach
-------------------------------
-
-When avoiding `mach` for some reason, one can run `servowpt.py`
-directly. However, this requires that all the dependencies for
-`wptrunner` are available in the current python environment.
 
 Running the tests manually
 --------------------------
